modified:   lab8/kern/fs/devs/dev.h 	modified:   lab8/kern/fs/swap/swapfs.c 	modified:   lab8/kern/mm/kmalloc.c 	modified:   lab8/kern/mm/pmm.c 	modified:   lab8/kern/process/proc.c 	modified:   lab8/kern/schedule/default_sched.c 	modified:   lab8/kern/sync/monitor.h 	modified:   lab8/kern/syscall/syscall.c 	modified:   lab8/kern/trap/trap.c 	modified:   lab8/user/libs/syscall.c 	modified:   lab8/user/libs/syscall.h

+++ lab8/user/libs/syscall.h
@@ -3,7 +3,7 @@
 
 int sys_exit(int64_t error_code);
 int sys_fork(void);
-int sys_wait(int64_t pid, int64_t *store);
+int sys_wait(int64_t pid, int *store);
 int sys_exec(const char *name, int64_t argc, const char **argv);
 int sys_yield(void);
 int sys_kill(int64_t pid);
@@ -30,4 +30,3 @@ void sys_lab6_set_priority(uint64_t priority); //only for lab6
 
 
 #endif /* !__USER_LIBS_SYSCALL_H__ */
-
